refactor(channels/telegram): extract tests to sibling file
Moves the inline #[cfg(test)] mod tests block out of channel.rs into
channel_tests.rs to bring the production file under the project's
<=500 line guideline. No behavior change.
